Updated to Version 2.1
Combat restricted to once per turn by adding a 2nd phase to turns.

+++ Assets/Movements/SharedMovement.cs
@@ -23,7 +23,7 @@ public class SharedMovement : MonoBehaviour
     //Units position on top of ground
     float stance = 0;
     //Whose turn it is
-    public static int turn = 0;
+    public static float turn = 0;
 
     /* 
         Used in error in ShowMovement
